day01_csv_sales/day01_csv_sales/requirements.txt
pandas
SQLAlchemy
psycopg2-binary
pyyaml
python-dotenv